requirements.txt
# Do NOT include azure-functions-worker (managed by Azure Functions)

azure-functions
azure-communication-email
